edge/ast-service/requirements.txt
numpy
librosa
torch
torchaudio
transformers
optimum[onnxruntime]
scikit-maad
pandas
psycopg2-binary
